backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-generativeai==0.3.0
httpx==0.25.1
redis==5.0.1
aioredis==2.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
gunicorn==21.2.0